Move ApiKey into propolis-datas
