+++ src/db.rs
@@ -8,8 +8,6 @@ use sqlx::{
 use std::env;
 use std::str::FromStr;
 
-#[cfg(feature = "with_predictions")]
-use crate::prediction::key::{ApiKey, ApiKeyStore};
 use crate::structs::{Statement, VoteHistoryItem};
 use crate::structs::{StatementStats, TargetSegment, User, Vote};
 
@@ -454,38 +452,4 @@ pub async fn get_followups(statement_id: i64, pool: &SqlitePool) -> Result<Vec<i
     )
     .fetch_all(pool)
     .await?)
-}
-
-#[cfg(feature = "with_predictions")]
-#[async_trait::async_trait]
-impl ApiKeyStore for SqlitePool {
-    async fn store(&mut self, item: &ApiKey) -> anyhow::Result<ApiKey> {
-        sqlx::query!(
-            "INSERT INTO api_keys (hash, note) VALUES (?, ?)",
-            item.hash,
-            item.note
-        )
-        .execute(self as &SqlitePool)
-        .await?;
-        let r = self.by_hash(item.hash.as_str()).await?;
-        r.ok_or(anyhow::anyhow!("Unable to retrieve just stored value"))
-    }
-    async fn by_id(&self, id: i64) -> anyhow::Result<Option<ApiKey>> {
-        Ok(sqlx::query_as!(
-            ApiKey,
-            "SELECT id, hash, note FROM api_keys WHERE id = ?",
-            id
-        )
-        .fetch_optional(self)
-        .await?)
-    }
-    async fn by_hash(&self, hash: &str) -> anyhow::Result<Option<ApiKey>> {
-        Ok(sqlx::query_as!(
-            ApiKey,
-            "SELECT id, hash, note FROM api_keys WHERE hash = ?",
-            hash
-        )
-        .fetch_optional(self)
-        .await?)
-    }
 }
